devtools/client/webide/test/test_newapp.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        (async function() {
          let win = await openWebIDE();
          let winProject = getProjectWindow(win);
          let tmpDir = FileUtils.getDir("TmpD", []);
          await winProject.projectList.newApp({
            index: 0,
            name: "webideTmpApp",
            folder: tmpDir
          });

          let project = win.AppManager.selectedProject;
          tmpDir = FileUtils.getDir("TmpD", ["webidetmpapp"]);
          ok(tmpDir.isDirectory(), "Directory created");
          is(project.location, tmpDir.path, "Location is valid (and lowercase)");
          is(project.name, "webideTmpApp", "name field has been updated");

          // Clean up
          tmpDir.remove(true);
          await closeWebIDE(win);
          await removeAllProjects();
          SimpleTest.finish();
        })();
      };
    </script>
  </body>
</html>
